bump deps and fix linter

+++ contracts/interfaces/IWETH.sol
@@ -5,9 +5,9 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IWETH is IERC20 {
-    event Deposit(address indexed dst, uint wad);
+    event Deposit(address indexed dst, uint256 wad);
 
-    event Withdrawal(address indexed src, uint wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
     function deposit() external payable;
